Add EBS CSI driver for EKS (#97)
* Add EBS CSI driver for EKS

Without the EBS CSI driver, the default `StorageClass` is no longer sufficient on recent EKS - persistent volumes cannot be provisioned. This change adds the EBS "Addon" for EKS and attaches the default AWS Policy for EBS CSI Driver  to the cluster nodegroup. Note: this just one of the ways it can be done, but others are not as straightforward to implement currently.

Ref: https://github.com/kubernetes-sigs/aws-ebs-csi-driver/blob/master/docs/install.md

* fix(addonName): its only working with addonName

Signed-off-by: Christopher Haar <christopher.haar@upbound.io>

---------

Signed-off-by: Christopher Haar <christopher.haar@upbound.io>
Co-authored-by: Christopher Haar <christopher.haar@upbound.io>